+++ CSharp/addons/YATI/TiledImport.cs
@@ -26,12 +26,12 @@ using Godot;
 namespace YATI;
 
 [Tool]
-public partial class TiledImport : EditorPlugin
+public partial class TiledImport : EditorPlugin 
 {
 	private EditorImportPlugin _xmlImport;
-	
+		
 	public override string _GetPluginName() => "Yet another Tiled importer";
-	
+		
 	public override void _EnterTree()
 	{
 		_xmlImport = new Importer();