+++ CSharp/runtime/TilesetCreator.cs
@@ -414,7 +414,7 @@ public class TilesetCreator
             TileData currentTile;
             if (tile.ContainsKey("image"))
             {
-                // Tile with it's own image -> separate atlas source
+                // Tile with its own image -> separate atlas source
                 _currentAtlasSource = new TileSetAtlasSource();
                 var addedSourceId = _tileset.AddSource(_currentAtlasSource, GetSpecialProperty(tile, GodotAtlasIdProperty));
                 RegisterAtlasSource(addedSourceId, 1, tileId, Vector2I.Zero);