+++ GDScript/addons/YATI/TilemapCreator.gd
@@ -67,6 +67,7 @@ var _add_class_as_metadata = false
 var _add_id_as_metadata = false
 var _dont_use_alternative_tiles = false
 var _custom_data_prefix: String = ""
+var _tileset_save_path: String = ""
 var _object_groups
 var _ct: CustomTypes = null
 
@@ -132,6 +133,10 @@ func set_custom_data_prefix(value: String):
 
 func set_custom_types(ct: CustomTypes):
 	_ct = ct
+
+
+func set_save_tileset_to(path: String):
+	_tileset_save_path = path
 
 
 func get_tileset():
@@ -221,6 +226,18 @@ func create(source_file: String):
 	if base_dictionary.has("layers"):
 		for layer in base_dictionary["layers"]:
 			handle_layer(layer, _base_node)
+
+	if _tileset_save_path != "":
+		var save_ret = ResourceSaver.save(_tileset, _tileset_save_path)
+		if save_ret == OK:
+			print("Successfully saved tileset to '" + _tileset_save_path + "'")
+			for node in _base_node.find_children("*", "TileMapLayer"):
+				if node is TileMapLayer and node.tile_set.resource_path == "":
+					node.tile_set = ResourceLoader.load(_tileset_save_path)
+		else:
+			printerr("Saving tileset returned error " + str(save_ret))
+			_error_count += 1
+
 
 	if base_dictionary.has("properties"):
 		handle_properties(_base_node, base_dictionary["properties"])
@@ -522,6 +539,7 @@ func create_map_from_data(layer_data: Array, offset_x: int, offset_y: int, map_w
 		var flipped_d = (int_id & FLIPPED_DIAGONALLY_FLAG) > 0
 		var gid: int = int_id & 0x0FFFFFFF
 		if gid <= 0: continue
+		@warning_ignore("integer_division")
 		var cell_coords = Vector2(cell_counter % map_width + offset_x, cell_counter / map_width + offset_y)
 
 		var source_id = get_matching_source_id(gid)
@@ -543,6 +561,7 @@ func create_map_from_data(layer_data: Array, offset_x: int, offset_y: int, map_w
 			if atlas_source.get_atlas_grid_size() == Vector2i.ONE:
 				atlas_coords = Vector2i.ZERO
 			else:
+				@warning_ignore("integer_division")
 				atlas_coords = Vector2(effective_gid % atlas_width, effective_gid / atlas_width)
 		if not atlas_source.has_tile(atlas_coords):
 			atlas_source.create_tile(atlas_coords)
@@ -597,7 +616,7 @@ func create_map_from_data(layer_data: Array, offset_x: int, offset_y: int, map_w
 
 func get_godot_type(godot_type_string: String):
 	var gts = godot_type_string.to_lower()
-	var _godot_type = {
+	var godot_type = {
 		"": _godot_type.EMPTY,
 		"collision": _godot_type.BODY,
 		"staticbody": _godot_type.BODY,
@@ -611,7 +630,7 @@ func get_godot_type(godot_type_string: String):
 		"polygon": _godot_type.POLYGON,
 		"instance": _godot_type.INSTANCE
 	}.get(gts, _godot_type.UNKNOWN)
-	return _godot_type
+	return godot_type
 
 
 func get_godot_node_type_property(obj: Dictionary):
@@ -834,6 +853,7 @@ func handle_object(obj: Dictionary, layer_node: Node, tileset: TileSet, offset: 
 			if atlas_width <= 0: return
 
 			var effective_gid: int = gid - _first_gids[get_first_gid_index(gid)]
+			@warning_ignore("integer_division")
 			var atlas_coords = Vector2(effective_gid % atlas_width, effective_gid / atlas_width)
 			if not gid_source.has_tile(atlas_coords):
 				gid_source.create_tile(atlas_coords)
@@ -964,8 +984,8 @@ func handle_object(obj: Dictionary, layer_node: Node, tileset: TileSet, offset: 
 		obj_text.visible = obj_visible
 		var txt = obj["text"]
 		obj_text.text = txt.get("text", "Hello World")
-		var wrap = txt.get("wrap", false)
-		obj_text.autowrap_mode = TextServer.AUTOWRAP_WORD_SMART if wrap else TextServer.AUTOWRAP_OFF
+		var wrap_ = txt.get("wrap", false)
+		obj_text.autowrap_mode = TextServer.AUTOWRAP_WORD_SMART if wrap_ else TextServer.AUTOWRAP_OFF
 		var align_h = txt.get("halign", "left")
 		match align_h:
 			"left": obj_text.horizontal_alignment = HORIZONTAL_ALIGNMENT_LEFT
@@ -1547,7 +1567,7 @@ func load_resource_from_file(path: String):
 	return ret
 	
 	
-func get_bitmask_integer_from_string(mask_string: String, max: int):
+func get_bitmask_integer_from_string(mask_string: String, max_len: int):
 	var ret: int = 0
 	var s1_arr = mask_string.split(",", false)
 	for s1 in s1_arr:
@@ -1557,12 +1577,12 @@ func get_bitmask_integer_from_string(mask_string: String, max: int):
 			var i2 = int(s2_arr[1]) if s2_arr[1].is_valid_int() else 0
 			if i1 == 0 or i2 == 0 or i1 > i2: continue
 			for i in range(i1, i2+1):
-				if i <= max:
-					ret += pow(2, i-1)
+				if i <= max_len:
+					ret += int(pow(2, i-1))
 		elif s1.is_valid_int():
 			var i = int(s1)
-			if i <= max:
-				ret += pow(2, i-1)
+			if i <= max_len:
+				ret += int(pow(2, i-1))
 	return ret
 
 
@@ -1651,6 +1671,8 @@ func handle_properties(target_node: Node, properties: Array):
 			target_node.use_parent_material = val.to_lower() == "true"
 	
 		# TileMapLayer properties
+		elif name.to_lower() == "tile_set" and type == "file" and target_node is TileMapLayer:
+			target_node.tile_set = load_resource_from_file(val)
 		elif name.to_lower() == "y_sort_origin" and type == "int" and target_node is TileMapLayer:
 			target_node.y_sort_origin = int(val)
 		elif name.to_lower() == "x_draw_order_reversed" and type == "bool" and target_node is TileMapLayer: